public List<Quest> QuestsAvailableHere { get; set; } = new List<Quest>(); added to Location Class
Added our first quest to the WorldFactory Class "Clear the Herbilist's garden" by using the new QuestAvalibaleHere property.

Added a new intermediary class to keep track of a quest the player has and their staus called QuestStaus.

In the Player class, we added a new property called Quests which is a List of QuestStatus items, becuase it is obervable we can automatically update the UI for quests when the Player recives a new Quest or compltes one of their existing ones.

In the gameSession class we add a new function called GivePLayerQuestsAtLocation which checks to see i fthe Player's Quests List has the Quests listed in the Location's Quest List. If the PLayer does not have a the Quests we add the Quests from the Location.

To display the Quests the Player has, we updated the MainWindow.xaml file to bind the data to the PlayerQuest property in the QuestStatus class, becuase PlayerQuest is of type Quest it also has a propty called Name which is what we access to diplay the name of the quest. isCompleted is also a property of QuestStatus.