Refactoring
In GameSession class replaced the getters with Lambda functions for all HasLocationTo properties.

in ItemFactory class, moved the line _standardGameItems = new List<GameItem>() from inside the constructor to the line where the variable declaration happens -line 13.

In World class AddLocation Constructor, we modified the ImageName variable to have the file location path so we don't have to type out the file path everytime we want to add a new location in the WorldFactory Class.

Same change as above made to the Monster Class so we are consistent with string concatonation type:
$"pack://application:,,,/Engine;component/Images/Monsters/{imageName}"

in MainWindow.xaml.cs we made the same changes as GameSession class, moving the varible declaration and creation to one line and making the variable readonly so we dont accidentally override it.