Player class changes: function to remove items from player inventory - RemoveItemsFromInventory function to check if player has required items to complete the quest - HassAllTheseItems
GamesSession class changes:
let player know they reviced the quest by raising messeges to display to UI
using GivePLayerQuestsAtLocation function.
modify the CurrentLocation setter, added a call to a new function – CompleteQuestsAtLocation(). So, when the player moves to a new location, this function will be called, and we will see if the player can complete the quests at that location.

CompletQuestsAtLocation function:
loop through all the quests available at this location, . If this location does not have any quests, we will loop zero times.
If “questToComplete” is not null (the player has this quest, and has not completed it), we check if the player has all the items needed to complete the quest (the “if” section, starting on line 140).

If the player has all the quest completion items, we remove them from the player’s inventory. Then, we give the player the rewards for completing the quest.

Finally,we mark the quest as completed.