Started Adding Traders
Created Trader class and TraderFactory Class.

Modified Location class to add the TraderHere property so traders can "live" at the location.

Modified WorldFactory, if Location has trader, the trader will be populated in the WorldFactory.

In GameSession ViewModel class
we update the UI to let it know if there is a trader at the player's current location. We add a new property called CurrentTrader and its backing Variable _currentTrader. CurrentTrader will be set when the player moves to a new location, we do this inside the CurrentLocation setter.
We also added a Boolean property called HasTrader, to use for the UI to decide whether or not to display the Trade button which will only appear if the location the player is at has a Trader - the CurrentTrader property.

In MainWindow.xaml we add the Trade Button to the UI, Its Visibility will be based on the Boolean HasTrader property in GameSession.cs.